Fxi/0.5 io reader

+++ skainet-io/skainet-io-gguf/src/commonMain/kotlin/sk/ainet/io/gguf/GGUFReader.kt
@@ -147,7 +147,7 @@ class GGUFReader(
         offs = newOffs
         val newAlign = fields["general.alignment"]
         if (newAlign != null && newAlign.types == listOf(GGUFValueType.UINT32)) {
-            alignment = newAlign.parts.last()[0] as Int
+            alignment = (newAlign.parts.last()[0] as UInt).toInt()
         }
         val padding = offs % alignment
         if (padding != 0) {
